backend-main/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
mysql-connector-python==9.2.0